Delete the oldest tracked version metadata files after commit
